microsoft365/outlook-mcp/pyproject.toml
[project]
name = "outlook-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.14",
    "azure-core>=1.35.0",
    "azure-identity>=1.23.0",
    "fastmcp>=2.10.4",
    "markdown>=3.8.2",
    "msgraph-sdk>=1.37.0",
    "typing-extensions>=4.14.1",
]
